added alerts to event listeners

+++ character/character.js
@@ -27,13 +27,16 @@ var allieEl = document.getElementById('choice-one');
 var allie = function(event){
   event.stopPropagation();
   event.preventDefault();
-  localStorage.setItem('maxHealth', '../images/IMG_0009.jpg');
-  localStorage.setItem('mediumHealth', '../images/IMG_0010.jpg');
-  localStorage.setItem('lowHealth', '../images/IMG_0011');
-  localStorage.setItem('health', 90);
-  localStorage.setItem('grade', 100);
-  localStorage.setItem('social', 120);
-  location.href = '../game/dayOne.html';
+  var check = confirm('do you wish to continue');
+  if(check === true){
+    localStorage.setItem('maxHealth', '../images/IMG_0009.jpg');
+    localStorage.setItem('mediumHealth', '../images/IMG_0010.jpg');
+    localStorage.setItem('lowHealth', '../images/IMG_0011');
+    localStorage.setItem('health', 90);
+    localStorage.setItem('grade', 100);
+    localStorage.setItem('social', 120);
+    location.href = '../game/dayOne.html';
+  }
 };
 allieEl.addEventListener('click', allie, false);
 //Logan's data
@@ -41,13 +44,16 @@ var loganEl = document.getElementById('choice-two');
 var logan = function(event){
   event.stopPropagation();
   event.preventDefault();
-  localStorage.setItem('maxHealth', '../images/IMG_0006.jpg');
-  localStorage.setItem('mediumHealth', '../images/IMG_0007.jpg');
-  localStorage.setItem('lowHealth', '../images/IMG_0008');
-  localStorage.setItem('health', 120);
-  localStorage.setItem('grade', 100);
-  localStorage.setItem('social', 80);
-  location.href = '../game/dayOne.html';
+  var check = confirm('do you wish to continue');
+  if(check === true){
+    localStorage.setItem('maxHealth', '../images/IMG_0006.jpg');
+    localStorage.setItem('mediumHealth', '../images/IMG_0007.jpg');
+    localStorage.setItem('lowHealth', '../images/IMG_0008');
+    localStorage.setItem('health', 120);
+    localStorage.setItem('grade', 100);
+    localStorage.setItem('social', 80);
+    location.href = '../game/dayOne.html';
+  }
 };
 loganEl.addEventListener('click', logan, false);
 //Luis' data
@@ -55,13 +61,16 @@ var luisEl = document.getElementById('choice-three');
 var luis = function(event){
   event.stopPropagation();
   event.preventDefault();
-  localStorage.setItem('maxHealth', '../images/IMG_0015.jpg');
-  localStorage.setItem('mediumHealth', '../images/IMG_0016.jpg');
-  localStorage.setItem('lowHealth', '../images/IMG_0017');
-  localStorage.setItem('health', 110);
-  localStorage.setItem('grade', 110);
-  localStorage.setItem('social', 80);
-  location.href = '../game/dayOne.html';
+  var check = confirm('do you wish to continue');
+  if(check === true){
+    localStorage.setItem('maxHealth', '../images/IMG_0015.jpg');
+    localStorage.setItem('mediumHealth', '../images/IMG_0016.jpg');
+    localStorage.setItem('lowHealth', '../images/IMG_0017');
+    localStorage.setItem('health', 110);
+    localStorage.setItem('grade', 110);
+    localStorage.setItem('social', 80);
+    location.href = '../game/dayOne.html';
+  }
 };
 luisEl.addEventListener('click', luis, false);
 //david's data
@@ -69,12 +78,15 @@ var davidEl = document.getElementById('choice-four');
 var david = function(event){
   event.stopPropagation();
   event.preventDefault();
-  localStorage.setItem('maxHealth', '../images/IMG_0012.jpg');
-  localStorage.setItem('mediumHealth', '../images/IMG_0013.jpg');
-  localStorage.setItem('lowHealth', '../images/IMG_0014');
-  localStorage.setItem('health', 80);
-  localStorage.setItem('grade', 100);
-  localStorage.setItem('social', 120);
-  location.href = '../game/dayOne.html';
+  var check = confirm('do you wish to continue');
+  if(check === true){
+    localStorage.setItem('maxHealth', '../images/IMG_0012.jpg');
+    localStorage.setItem('mediumHealth', '../images/IMG_0013.jpg');
+    localStorage.setItem('lowHealth', '../images/IMG_0014');
+    localStorage.setItem('health', 80);
+    localStorage.setItem('grade', 100);
+    localStorage.setItem('social', 120);
+    location.href = '../game/dayOne.html';
+  }
 };
 davidEl.addEventListener('click', david, false);
